implementations/python-sdk/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aether"
version = "0.1.0"
description = "Aether Python SDK"
requires-python = ">=3.10"
dependencies = [
  "PyNaCl",
  "blake3",
  "bech32",
  "websockets",
]

[project.optional-dependencies]
dev = [
  "black",
  "mypy",
  "pytest",
  "ruff",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["aether*"]
